+++ koni_rar/lib/src/rar_format.dart
@@ -28,7 +28,7 @@ final class RarFormat extends ArchiveFormat {
   }
 
   @override
-  Future<ArchiveReader> openReader(
+  Future<ArchiveReader> createReader(
     ByteSource source,
     ArchiveReadOptions options,
   ) async {